Updated docker-compose to expose correct ports. Updated dockerfile to correct build process. Added JsonProperty flags to security claims model to deserialize _id fields correctly. Updated Draft upload dir cleaner to mkdir if dir isnt found. Changed healthcheck controller route.

+++ Template-WebAPI/BackgroundServices/TemplateDraftUploadDirCleaner.cs
@@ -10,17 +10,24 @@ namespace Template_WebAPI
   {
     public void ClearDraftDir()
     {
-      var directoryName = Path.Combine("Resources", "File");
-      var pathToDirectory = Path.Combine(Directory.GetCurrentDirectory(), directoryName);
-      string[] files = Directory.GetFiles(pathToDirectory);
 
-      foreach (string file in files)
+      var resourcesDir = Path.Combine(Directory.GetCurrentDirectory(), "Resources");
+      var pathToDirectory = Path.Combine(resourcesDir, "File");
+
+      DirectoryInfo di = Directory.CreateDirectory(resourcesDir);
+
+      if (Directory.Exists(resourcesDir) && Directory.Exists(pathToDirectory))
       {
-        FileInfo fileInfo = new FileInfo(file);
+        string[] files = Directory.GetFiles(pathToDirectory);
 
-        if (fileInfo.CreationTime < DateTime.Now.AddMinutes(-60))
+        foreach (string file in files)
         {
-          fileInfo.Delete();
+          FileInfo fileInfo = new FileInfo(file);
+
+          if (fileInfo.CreationTime < DateTime.Now.AddMinutes(-60))
+          {
+            fileInfo.Delete();
+          }
         }
       }
     }
